requirements.txt
attr==0.3.2
ConfigParser==7.0.0
contextlib2==21.6.0
cryptography==42.0.5
Cython==3.0.10
dl==0.1.0
docutils==0.21.1
filelock==3.13.4
fitz==0.0.1.dev2
HTMLParser==0.0.2
ipython==8.12.3
ipywidgets==8.1.2
Jinja2==3.1.3
jnius==1.1.0
keyring==25.1.0
loguru==0.7.2
mock==5.1.0
numba==0.59.0
numpy==1.26.4
openpyxl==3.1.2
Pillow==10.3.0
prettytable==3.10.0
protobuf==5.26.1
psycopg2==2.9.9
pymorphy2==0.9.1
pyOpenSSL==24.1.0
pytesseract==0.3.10
pywin32==306
railroad==0.5.0
redis==5.0.3
schedule==1.2.1
Sphinx==7.3.7
thread==1.1.1
tornado==6.4
tqdm==4.66.2
trove_classifiers==2024.4.10
urllib3_secure_extra==0.1.0
vkbottle==4.3.12
xmlrpclib==1.0.1
